Readme
This is a collection of python methods that will enable one to apply machine learning methods in python in the differentiation of different proteomic or metabolomic samples. The investigations and methods described in this repository are cumulative efforts in the extensive and complete comparable characterization of mass spectrometry-based samples.  Since these methods are exclusively applicable to LC-MS data, it should be noted that most of the differences that can be observed by machine learning algorithms are not human observable and should be treated as such. This repository is simply a compilation of jupyter notebooks that have code that is useful when one is tackling the mentioned problem. It should be noted that sequence of steps and algorithms used is dependent on personal choice and cannot be regarded as final or superior in any form.

Any ideas/suggestions are always welcome.